+++ OpenSim/Moco/MocoControlBoundConstraint.h
@@ -25,7 +25,7 @@ namespace OpenSim {
 
 class MocoProblemInfo;
 
-/** This class constrains any number of control signals from ScalarActautor%s to
+/** This class constrains any number of control signals from ScalarActuator%s to
 be between two time-based functions. It is possible to constrain the control
 signal to be exactly to a provided function; see the equality_with_lower
 property.
@@ -34,10 +34,15 @@ If a function is a GCVSpline, we ensure that the spline covers the entire
 possible time range in the problem (using the problem's time bounds). We do
 not perform such a check for other types of functions.
 
+If you wish to constrain all control signals except those associated with a
+user-defined controller (e.g., PrescribedController), pass 'true' to
+`setIgnoreControlledActuators()`.
+
 @note If you omit the lower and upper bounds, then this class will not
 constrain any control signals, even if you have provided control paths.
 
 @note This class can only constrain control signals for ScalarActuator%s.
+
 @ingroup mocopathcon */
 class OSIMMOCO_API MocoControlBoundConstraint : public MocoPathConstraint {
     OpenSim_DECLARE_CONCRETE_OBJECT(
@@ -84,6 +89,16 @@ public:
     //// @copydoc setEqualityWithLower()
     bool getEqualityWithLower() const { return get_equality_with_lower(); }
 
+    /// If true, do not constrain controls associated with user-defined
+    /// controllers.
+    void setIgnoreControlledActuators(bool v) {
+        set_ignore_controlled_actuators(v);
+    }
+    /// @copydoc setIgnoreControlledActuators()
+    bool getIgnoreControlledActuators() const {
+        return get_ignore_controlled_actuators();
+    }
+
 protected:
     void initializeOnModelImpl(
             const Model& model, const MocoProblemInfo&) const override;
@@ -102,6 +117,9 @@ private:
     OpenSim_DECLARE_PROPERTY(equality_with_lower, bool,
             "The control must be equal to the lower bound; "
             "upper must be unspecified (default: false).");
+    OpenSim_DECLARE_PROPERTY(ignore_controlled_actuators, bool,
+            "If true, do not constrain controls belonging to actuators "
+            "controlled by user-defined controllers (default: false).");
 
     void constructProperties();